src/bot/requirements.txt
datasets>=4.4.0
numpy>=2.2.0
pyarrow>=12.0.0
python-chess>=1.999
torch>=2.3.0
tqdm>=4.66.0
